BugFix: i). Now if user already exist then he cann't enter the detail again ii). UpdateProfile empty entry update fixed

+++ app/src/main/java/com/example/xms/UserDashboardActivity.java
@@ -1,8 +1,13 @@
 package com.example.xms;
 
+import android.Manifest;
 import android.app.Activity;
+import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.Uri;
+import android.net.wifi.WifiInfo;
+import android.net.wifi.WifiManager;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -10,12 +15,15 @@ import android.widget.Button;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.DocumentId;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -39,7 +47,10 @@ public class UserDashboardActivity extends Activity {
     String timestamp;
     FirebaseFirestore fbd;
     FirebaseUser user;
-//    DocumentReference documentReference;
+    DocumentReference persDetail;
+
+
+    public static final int requestCode = 2019123;
 
     @Override
     protected void onCreate(Bundle savedInstanceState){
@@ -63,18 +74,6 @@ public class UserDashboardActivity extends Activity {
         fbd = FirebaseFirestore.getInstance();
         user = FirebaseAuth.getInstance().getCurrentUser();
 
-//        documentReference = fbd.collection("users").document(user.getUid().toString().trim());
-//
-//        documentReference.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
-//            @Override
-//            public void onSuccess(DocumentSnapshot documentSnapshot) {
-//                fname = documentSnapshot.get("firstname").toString().trim();
-//                lname = documentSnapshot.get("lastname").toString().trim();
-//                pNum = documentSnapshot.get("phone").toString().trim();
-//            }
-//        });
-
-
         profile.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -83,23 +82,39 @@ public class UserDashboardActivity extends Activity {
             }
         });
 
+
+
+//        ActivityCompat.shouldShowRequestPermissionRationale(UserDashboardActivity.this, Manifest.permission.ACCESS_WIFI_STATE);
+
+        if( ContextCompat.checkSelfPermission(this,Manifest.permission.ACCESS_COARSE_LOCATION)!= PackageManager.PERMISSION_GRANTED) {
+            requestPermissions(new String[] {Manifest.permission.ACCESS_COARSE_LOCATION},requestCode);
+        }
+
+//        WifiManager wifiMgr = (WifiManager) getApplicationContext().getSystemService(Context.WIFI_SERVICE);                                  // TODO: Resolve the incosistent BSSID info
+//        WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
+//        Toast.makeText(UserDashboardActivity.this, wifiInfo.toString()  ,Toast.LENGTH_LONG).show();
+
+
         entry.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                timestamp = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss").format(new Date());
 
-                Map<String,String> detail = new HashMap<>();
-                detail.put(timestamp.trim(),"Entry");
-                fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
-
-//                Map<String,String> personPresent = new HashMap<>();
-//
-//                personPresent.put(fname + " " + lname, pNum + " " + user.getEmail() );
-
-                fbd.collection("personPresent").document(user.getUid().trim()).set(detail);
-
-                Toast.makeText(UserDashboardActivity.this,"Entry Detail Recorded",Toast.LENGTH_SHORT).show();
-
+                persDetail = fbd.collection("personPresent").document(user.getUid().trim());
+                persDetail.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                    @Override
+                    public void onSuccess(DocumentSnapshot documentSnapshot) {
+                        if(documentSnapshot.exists()){
+                            Toast.makeText(UserDashboardActivity.this,"Already Inside",Toast.LENGTH_SHORT).show();
+                        }else{
+                            timestamp = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss").format(new Date());
+                            Map<String,String> detail = new HashMap<>();
+                            detail.put(timestamp.trim(),"Entry");
+                            fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
+                            fbd.collection("personPresent").document(user.getUid().trim()).set(detail);
+                            Toast.makeText(UserDashboardActivity.this,"Entry Detail Recorded",Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
 
             }
         });
@@ -107,17 +122,24 @@ public class UserDashboardActivity extends Activity {
         exit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                timestamp = new SimpleDateFormat("dd-MM-yyyy kk:mm:ss").format(new Date());
 
 
-                Map<String,String> detail = new HashMap<>();
-                detail.put(timestamp.trim(),"Exit");
-                fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
-
-                fbd.collection("personPresent").document(user.getUid().trim()).delete();
-
-                Toast.makeText(UserDashboardActivity.this,"Exit Detail Recorded",Toast.LENGTH_SHORT).show();
-
+                persDetail = fbd.collection("personPresent").document(user.getUid().trim());
+                persDetail.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                    @Override
+                    public void onSuccess(DocumentSnapshot documentSnapshot) {
+                        if(documentSnapshot.exists()){
+                            timestamp = new SimpleDateFormat("dd-MM-yyyy kk:mm:ss").format(new Date());
+                            Map<String,String> detail = new HashMap<>();
+                            detail.put(timestamp.trim(),"Exit");
+                            fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
+                            fbd.collection("personPresent").document(user.getUid().trim()).delete();
+                            Toast.makeText(UserDashboardActivity.this,"Exit Detail Recorded",Toast.LENGTH_SHORT).show();
+                        }else{
+                            Toast.makeText(UserDashboardActivity.this,"Already Outside",Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
 
             }
         });
